requirements.txt
kubernetes==12.0.1
pbr==5.5.1
PyYAML==5.3.1